+++ tests/Unit/Scopes/Classic/VerifiedAtScopeTest.php
@@ -13,10 +13,10 @@ declare(strict_types=1);
 
 namespace Cog\Tests\Flag\Unit\Scopes\Classic;
 
-use Carbon\Carbon;
 use Cog\Tests\Flag\Stubs\Models\Classic\EntityWithVerifiedAt;
 use Cog\Tests\Flag\Stubs\Models\Classic\EntityWithVerifiedAtUnapplied;
 use Cog\Tests\Flag\TestCase;
+use Illuminate\Support\Carbon;
 
 class VerifiedAtScopeTest extends TestCase
 {